destroyアクションの条件分岐

+++ app/controllers/products_controller.rb
@@ -36,8 +36,12 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if product.user.id == current_user.id
+      product.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
